+++ src/Adapter/Filesystem/Fold.php
@@ -59,18 +59,13 @@ final class Fold
         }
 
         if ($specification instanceof EntitySpecification) {
-            $filter = $this->filter($specification);
+            $filter = $this->child($specification->specification());
 
             return static fn(Aggregate $aggregate) => $aggregate
                 ->entities()
                 ->find(static fn($entity) => $entity->name() === $specification->entity())
-                ->flatMap(
-                    static fn($entity) => $entity
-                        ->properties()
-                        ->find(static fn($property) => $property->name() === $specification->property()),
-                )
                 ->match(
-                    static fn($property) => $filter($property->value()),
+                    static fn($entity) => $filter($entity),
                     static fn() => false,
                 );
         }
@@ -150,14 +145,14 @@ final class Fold
     }
 
     /**
-     * @return callable(Aggregate\Collection\Entity): bool
+     * @return callable(Aggregate\Collection\Entity|Aggregate\Entity): bool
      */
     private function child(Specification $specification): callable
     {
         if ($specification instanceof Not) {
             $filter = $this->child($specification->specification());
 
-            return static fn(Aggregate\Collection\Entity $entity) => !$filter($entity);
+            return static fn(Aggregate\Collection\Entity|Aggregate\Entity $entity) => !$filter($entity);
         }
 
         if ($specification instanceof Composite) {
@@ -165,8 +160,8 @@ final class Fold
             $right = $this->child($specification->right());
 
             return match ($specification->operator()) {
-                Operator::and => static fn(Aggregate\Collection\Entity $entity) => $left($entity) && $right($entity),
-                Operator::or => static fn(Aggregate\Collection\Entity $entity) => $left($entity) || $right($entity),
+                Operator::and => static fn(Aggregate\Collection\Entity|Aggregate\Entity $entity) => $left($entity) && $right($entity),
+                Operator::or => static fn(Aggregate\Collection\Entity|Aggregate\Entity $entity) => $left($entity) || $right($entity),
             };
         }
 
@@ -178,7 +173,7 @@ final class Fold
 
         $filter = $this->filter($specification);
 
-        return static fn(Aggregate\Collection\Entity $entity) => $entity
+        return static fn(Aggregate\Collection\Entity|Aggregate\Entity $entity) => $entity
             ->properties()
             ->find(static fn($property) => $property->name() === $specification->property())
             ->match(